use usize in truncate

+++ tee/kernel/src/fs/fd.rs
@@ -242,7 +242,7 @@ pub trait OpenFileDescription: Send + Sync + 'static {
         Err(Error::inval(()))
     }
 
-    fn truncate(&self, length: u64) -> Result<()> {
+    fn truncate(&self, length: usize) -> Result<()> {
         let _ = length;
         Err(Error::inval(()))
     }
